Split test files
Why:

- Increase visibility through smaller files

This change addresses the need by:

- Creating multiple test files and splitting the tests between them
